Move every styled-component to /design-system
