+++ components/Loading/index.jsx
@@ -1,0 +1,7 @@
+import LoadingAnimation from "@/design-system/StyledLoading";
+
+function Loading() {
+  return <LoadingAnimation />;
+}
+
+export default Loading;
