+++ design-system/StyledExpenseDetail.js
