+++ design-system/StyledExpenseListDetail.js
